+++ framework/Source/CPTGradient.h
@@ -46,11 +46,26 @@ CPTGradientType;
     CGFunctionRef gradientFunction;
     CGFloat angle; // angle in degrees
     CPTGradientType gradientType;
+    CGPoint startAnchor;
+    CGPoint endAnchor;
 }
 
+/// @name Gradient Type
+/// @{
 @property (nonatomic, readonly, assign) CPTGradientBlendingMode blendingMode;
-@property (nonatomic, readwrite, assign) CGFloat angle;
 @property (nonatomic, readwrite, assign) CPTGradientType gradientType;
+/// @}
+
+/// @name Axial Gradients
+/// @{
+@property (nonatomic, readwrite, assign) CGFloat angle;
+/// @}
+
+/// @name Radial Gradients
+/// @{
+@property (nonatomic, readwrite, assign) CGPoint startAnchor;
+@property (nonatomic, readwrite, assign) CGPoint endAnchor;
+/// @}
 
 /// @name Factory Methods
 /// @{